# perf: Optimize dataobj.Reader allocs: Use bufpool and discard skipped values