**What this PR does / why we need it**:
This PR is attempting to reduce the memory consumption & execution time of reading from data objects.

I observed via benchmarking that currently we allocate every value we want to check predicates for, so we'll separately allocate every traceID when reading the traceID column in a predicate. For rows which match the predicate, we'll then fill in the other columns. I also added a step to discard columns without allocating if we are skipping through them - which happens during fill

I attempted to reduce the allocations here by allocating from a pool while we check predicates, then if the predicates are passing I permanently allocate them in a new, correctly sized, slice and return the values.
(I had a method which just used a []byte pool directly which was a bit faster, but it didn't support variably sized slices so I switched to bufpool which is slightly slower)

This method is effective at reducing allocations, but the overall runtime increased by 10-20%, so I'm conflicted if this is worth merging. It should generate less CPU load in the longer term due to less GC...

LogQL Benchmark results
```
goos: darwin
goarch: arm64
pkg: github.com/grafana/loki/v3/pkg/logql/bench
cpu: Apple M3 Max
                                                                                                    │ before_bench.txt │          after_bench.txt           │
                                                                                                    │      sec/op      │   sec/op     vs base               │
LogQL/dataobj/{region="ap-southeast-1"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]-14        28.19 ± 36%   26.70 ± 12%       ~ (p=0.631 n=10)
LogQL/dataobj/{region="ap-southeast-1"}_|~_"error|exception"_|_detected_level="error"_[FORWARD]-14         20.22 ± 43%   24.93 ± 19%       ~ (p=0.190 n=10)
geomean                                                                                                    23.87         25.80        +8.09%

                                                                                                    │ before_bench.txt │            after_bench.txt            │
                                                                                                    │       B/op       │     B/op       vs base                │
LogQL/dataobj/{region="ap-southeast-1"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]-14      4.709Gi ± 17%   3.305Gi ± 40%  -29.80% (p=0.005 n=10)
LogQL/dataobj/{region="ap-southeast-1"}_|~_"error|exception"_|_detected_level="error"_[FORWARD]-14       4.566Gi ±  1%   3.072Gi ± 10%  -32.71% (p=0.000 n=10)
geomean                                                                                                  4.637Gi         3.187Gi        -31.27%

                                                                                                    │ before_bench.txt │           after_bench.txt           │
                                                                                                    │    allocs/op     │  allocs/op   vs base                │
LogQL/dataobj/{region="ap-southeast-1"}_|~_"error|exception"_|_detected_level="error"_[BACKWARD]-14       104.72M ± 0%   16.72M ± 7%  -84.04% (p=0.000 n=10)
LogQL/dataobj/{region="ap-southeast-1"}_|~_"error|exception"_|_detected_level="error"_[FORWARD]-14        104.71M ± 0%   16.65M ± 4%  -84.10% (p=0.000 n=10)
geomean                                                                                                    104.7M        16.68M       -84.07%
```

Logs Reader Benchmark results
```
goos: darwin
goarch: arm64
pkg: github.com/grafana/loki/v3/pkg/dataobj
cpu: Apple M3 Max
              │ before_logs_reader.txt │    after_logs_reader_bufpool.txt    │
              │         sec/op         │   sec/op     vs base                │
LogsReader-14              2.366m ± 1%   2.878m ± 2%  +21.61% (p=0.000 n=10)

              │ before_logs_reader.txt │     after_logs_reader_bufpool.txt     │
              │          B/op          │     B/op       vs base                │
LogsReader-14           1027.7Ki ± 17%   607.1Ki ± 30%  -40.93% (p=0.000 n=10)

              │ before_logs_reader.txt │   after_logs_reader_bufpool.txt    │
              │       allocs/op        │ allocs/op   vs base                │
LogsReader-14             49300.0 ± 0%   889.0 ± 0%  -98.20% (p=0.000 n=10)
```